pyproject.toml
[project]
name = "live-swe-agent-leaderboard"
version = "1.0.0"
description = "Leaderboard for live-swe-agent scaffold on SWE-bench"
requires-python = ">=3.8"
dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "markdown>=3.4.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
